+++ CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to the "allyzio" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-### 0.0.1
+### 0.1.1
 
-- Initial version with the option to refactor your code
+- remove prompt refactor code. It was generating bugs
+- bugfix: refactor code
 
 ### 0.1.0
 
 - change menu
 - add option comment code
+
+### 0.0.1
+
+- Initial version with the option to refactor your code